refactor: consolidate release tooling and desktop/server helpers
- Extract scripts/lib/workspace-packages.mjs as the single source of truth for
  the workspace inventory and the root-version equality check; use it from
  verify-release-tag.mjs, pack-npm.mjs, and verify-npm-tarballs.mjs (validated
  end-to-end via a full pack + isolated install smoke).
- Add shared host predicates (isLoopbackHost, isWildcardHost, isPrivateIpv4,
  isVirtualInterface) and proxy-environment helpers (firstEnvValue,
  mergeNoProxySources, mergeNoProxy, loopbackNoProxyEntries) in
  workbench-server/src/infrastructure/network and import them from the desktop
  shell, removing the duplicated copies in main.ts, electron-download-env.ts,
  and daemon/urls.ts. isLoopbackHost now lowercases (matching the desktop
  behavior) so it also accepts case variants at the daemon bind check.
- desktop-shell logging.ts uses the server's resolveDataDir instead of a
  private copy; daemon/profile.ts keeps its injectable-home resolveDaemonPaths
  (covered by tests).
