+++ packages/workbench-server/src/index.ts
@@ -5,6 +5,7 @@ export {
 } from "./app/orchestrator-state.js";
 export { createApp } from "./app/server.js";
 export { version } from "./app/version.js";
+export * from "./infrastructure/network/index.js";
 export {
   inspectWorkbenchHome,
   type LegacyCredentialMigrationStatus,
