+++ scripts/verify-release-tag.mjs
@@ -1,18 +1,4 @@
-import { readFile } from "node:fs/promises";
-import { dirname, join, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
-
-const repoRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
-const packagePaths = [
-  "package.json",
-  join("packages", "contracts", "package.json"),
-  join("packages", "protocol", "package.json"),
-  join("packages", "harness", "package.json"),
-  join("packages", "tools", "package.json"),
-  join("packages", "workbench-server", "package.json"),
-  join("packages", "workbench-app", "package.json"),
-  join("packages", "desktop-shell", "package.json"),
-];
+import { assertWorkspaceVersionsMatch } from "./lib/workspace-packages.mjs";
 
 const args = process.argv.slice(2).filter((arg) => arg !== "--");
 const rawTag = args[0] ?? process.env.GITHUB_REF_NAME ?? "";
@@ -23,27 +9,7 @@ if (!tag) {
   );
 }
 
-const versions = new Map();
-for (const relativePath of packagePaths) {
-  const packageJson = await readJson(join(repoRoot, relativePath));
-  if (typeof packageJson.version !== "string" || !packageJson.version) {
-    throw new Error(`${relativePath} does not declare a version.`);
-  }
-  versions.set(relativePath, packageJson.version);
-}
-
-const rootVersion = versions.get("package.json");
-const mismatches = [...versions.entries()].filter(
-  ([, version]) => version !== rootVersion,
-);
-if (mismatches.length > 0) {
-  throw new Error(
-    `Workspace package versions must match root version ${rootVersion}:\n${mismatches
-      .map(([path, version]) => `  ${path}: ${version}`)
-      .join("\n")}`,
-  );
-}
-
+const rootVersion = await assertWorkspaceVersionsMatch();
 const expectedTag = `v${rootVersion}`;
 if (tag !== expectedTag) {
   throw new Error(
@@ -52,7 +18,3 @@ if (tag !== expectedTag) {
 }
 
 console.log(`Release tag ${tag} matches workspace version ${rootVersion}.`);
-
-async function readJson(path) {
-  return JSON.parse(await readFile(path, "utf8"));
-}
